+++ Source/Promise+map.swift
@@ -39,3 +39,30 @@ public func map<A, B, S: Sequence>(on queue: DispatchQueue = .main, series items
         }
     }
 }
+
+extension Promise where Result: Sequence {
+
+    /// Identical to `map()`, but can be chained on an existing promise that resolves to a sequence type
+    ///
+    /// - parameter queue:     dispatch queue to run the handler on
+    /// - parameter transform: transform function to run on each item
+    ///
+    /// - returns: Promise
+    public func map<B>(on queue: DispatchQueue = .main, _ transform: @escaping (Result.Iterator.Element) throws -> Promise<B>) -> Promise<[B]> {
+        return then(on: queue) { results in
+            return Bluebird.map(on: queue, results, transform)
+        }
+    }
+
+    /// Identical to `map(series)`, but can be chained on an existing promise that resolves to a sequence type
+    ///
+    /// - parameter queue:     dispatch queue to run the handler on
+    /// - parameter transform: transform function to run on each item
+    ///
+    /// - returns: Promise
+    public func mapSeries<B>(on queue: DispatchQueue = .main, _ transform: @escaping (Result.Iterator.Element) throws -> Promise<B>) -> Promise<[B]> {
+        return then(on: queue) { results in
+            return Bluebird.map(on: queue, series: results, transform)
+        }
+    }
+}